Worked on fixing auto-logout cookie bug
