+++ server/env.template
@@ -1,0 +1,11 @@
+# Mysql Database
+DATABASE_HOST="127.0.0.1"
+DATABASE_USER="root"
+DATABASE_PASSWORD="password"
+DATABASE_NAME="loginSystem"
+
+# Server port
+SERVER_PORT=3001
+
+# Session settings
+SESSION_SECRETS="SECRET_1 SECRET_2 SECRET_3 SECRET_4 ..."
